Music Converter Microservice/remusic-conversion/requirements.txt
tensorflow==2.11.0
keras==2.11.0
oemer[tf]
kafka-python
requests
music21
